requirements/base.txt
# Core Django
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Authentication
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9

# Celery
celery>=5.3.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# Redis
redis>=5.0.0
django-redis>=5.4.0

# Payment Gateways
stripe>=7.0.0
requests>=2.31.0

# Utilities
python-decouple>=3.8
Pillow>=10.0.0

# Filtering
django-filter>=23.3

# Environment
python-dotenv>=1.0.0
